+++ native/py_bindings.h
@@ -128,6 +128,9 @@ PyObject* mod_ycd_decode_frame_channels(PyObject*, PyObject* args);
 PyObject* mod_ycd_encode_frame_channels(PyObject*, PyObject* args);
 PyObject* mod_ycd_decode_quantized_values(PyObject*, PyObject* args);
 PyObject* mod_ycd_decode_linear_values(PyObject*, PyObject* args);
+PyObject* mod_ycd_track_sampler_new(PyObject*, PyObject* args);
+PyObject* mod_ycd_track_sampler_window(PyObject*, PyObject* args);
+PyObject* mod_ycd_track_sampler_retained_count(PyObject*, PyObject* args);
 PyObject* mod_vector_interpolate_many(PyObject*, PyObject* args);
 
 extern PyMethodDef module_methods[];